deployment/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Pillow==10.0.1
qrcode[pil]==7.4.2
pyzbar==0.1.9
Werkzeug==2.3.7
gunicorn==21.2.0

